+ AdamW beta1, beta2, eps to SFT CLI. + Ckpt'ing for vLLM hot-reload during evaluation. + Tests to validate model ckpt'ing and vllm hot-reload. + pytest-asyncio to project dev deps + Log tokens per second metrics + Type stubs for wandb so mypy is happy

+++ grpo_gsm8k/log_generations.py
@@ -97,7 +97,7 @@ def log_generations(
     # out.scores is a list of length T_gen with tensors of shape (B, V).
     if len(out.scores) > 0:
         logits_bt = torch.stack(out.scores, dim=1)  # (B, T_gen, V)
-        ent_bt = compute_entropy(logits_bt)  # (B, T_gen)
+        ent_bt = compute_entropy(logits_bt).to(torch.float32)  # (B, T_gen)
     else:
         ent_bt = torch.zeros((batch_size, 0), dtype=torch.float32, device=device)
